+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_models.py
@@ -3268,7 +3268,7 @@ class DocumentTableCell:
         )
 
 
-class ModelOperationSummary:
+class DocumentModelOperationSummary:
     """Model operation information, including the kind and status of the operation, when it was
     created, and more.
 
@@ -3314,14 +3314,14 @@ class ModelOperationSummary:
 
     def __repr__(self):
         return (
-            f"ModelOperationSummary(operation_id={self.operation_id}, status={self.status}, "
+            f"DocumentModelOperationSummary(operation_id={self.operation_id}, status={self.status}, "
             f"percent_completed={self.percent_completed}, created_on={self.created_on}, "
             f"last_updated_on={self.last_updated_on}, kind={self.kind}, "
             f"resource_location={self.resource_location}, api_version={self.api_version}, tags={self.tags})"
         )
 
     def to_dict(self) -> dict:
-        """Returns a dict representation of ModelOperationSummary.
+        """Returns a dict representation of DocumentModelOperationSummary.
 
         :return: dict
         :rtype: dict
@@ -3339,12 +3339,12 @@ class ModelOperationSummary:
         }
 
     @classmethod
-    def from_dict(cls, data: dict) -> "ModelOperationSummary":
-        """Converts a dict in the shape of a ModelOperationSummary to the model itself.
+    def from_dict(cls, data: dict) -> "DocumentModelOperationSummary":
+        """Converts a dict in the shape of a DocumentModelOperationSummary to the model itself.
 
-        :param dict data: A dictionary in the shape of ModelOperationSummary.
-        :return: ModelOperationSummary
-        :rtype: ModelOperationSummary
+        :param dict data: A dictionary in the shape of DocumentModelOperationSummary.
+        :return: DocumentModelOperationSummary
+        :rtype: DocumentModelOperationSummary
         """
         return cls(
             operation_id=data.get("operation_id", None),
@@ -3373,8 +3373,8 @@ class ModelOperationSummary:
         )
 
 
-class ModelOperationDetails(ModelOperationSummary):
-    """ModelOperationDetails consists of information about the model operation, including the result or
+class DocumentModelOperationDetails(DocumentModelOperationSummary):
+    """DocumentModelOperationDetails consists of information about the model operation, including the result or
     error of the operation if it has completed.
 
     Note that operation information only persists for 24 hours. If the operation was successful,
@@ -3419,7 +3419,7 @@ class ModelOperationDetails(ModelOperationSummary):
 
     def __repr__(self):
         return (
-            f"ModelOperationDetails(operation_id={self.operation_id}, status={self.status}, "
+            f"DocumentModelOperationDetails(operation_id={self.operation_id}, status={self.status}, "
             f"percent_completed={self.percent_completed}, created_on={self.created_on}, "
             f"last_updated_on={self.last_updated_on}, kind={self.kind}, "
             f"resource_location={self.resource_location}, result={repr(self.result)}, "
@@ -3427,7 +3427,7 @@ class ModelOperationDetails(ModelOperationSummary):
         )
 
     def to_dict(self) -> dict:
-        """Returns a dict representation of ModelOperationDetails.
+        """Returns a dict representation of DocumentModelOperationDetails.
 
         :return: dict
         :rtype: dict
@@ -3447,12 +3447,12 @@ class ModelOperationDetails(ModelOperationSummary):
         }
 
     @classmethod
-    def from_dict(cls, data: dict) -> "ModelOperationDetails":
-        """Converts a dict in the shape of a ModelOperationDetails to the model itself.
+    def from_dict(cls, data: dict) -> "DocumentModelOperationDetails":
+        """Converts a dict in the shape of a DocumentModelOperationDetails to the model itself.
 
-        :param dict data: A dictionary in the shape of ModelOperationDetails.
-        :return: ModelOperationDetails
-        :rtype: ModelOperationDetails
+        :param dict data: A dictionary in the shape of DocumentModelOperationDetails.
+        :return: DocumentModelOperationDetails
+        :rtype: DocumentModelOperationDetails
         """
         return cls(
             operation_id=data.get("operation_id", None),